Adjust footer elements and image size for mobile
As pointed out by my Devslopes mentor, the layout could have flowed a
little better in certain areas for tablet and smart phone sizes.

Keeping the #social-list in a row for tablet decreases the amount of
blank space on the page. Same goes with making the images bigger.

I also fixed the features images so ththey wouldn't stretch
so far and become visually unreadable.